Add files via upload
Not sure why the nav bar is showing up with a white border around it. It has to be something in the css. I've looked at it for too long so I need a break. The layout and css for everything else on the page is finalized. If y'all have time to put it into the web2py this week or weekend thats cool. Just don't freakout if you see the border around the nav bar. I know its there lol. Thanks team